feat(auth): add keycloak auth interceptor
This pull request adds a KeycloakAuthInterceptor on top of `got` hooks. Hooks
are supported since version 9 of `got` and they allow to use async functions
to manipulate the request. This pull request adds this functionality without
loosing backward compatibility. It also serves as an example of how authentication
against third party authentication providers can be implemented with hooks.

The complete code was updated to work with the latest version of `got`. The new
functionality is covered with tests.